[Refactor] Move data sub components
refs #79677
