+++ src/app/dmp/specify-data/data-mc/data-mc.component.spec.ts
@@ -1,7 +1,7 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {DataMcComponent} from './data-mc.component';
-import {TranslateTestingModule} from '../../../../../testing/translate-testing/translate-testing.module';
+import {TranslateTestingModule} from '../../../testing/translate-testing/translate-testing.module';
 
 describe('DataMcComponent', () => {
   let component: DataMcComponent;
