+++ src/app/pipe/dataset-source/dataset-source.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {DatasetSourcePipe} from '../dataset-source.pipe';
+import {DatasetSourcePipe} from './dataset-source.pipe';
 
 
 @NgModule({
